refactor : move winner test code to RacingCarsTest
